v1.3
-Update tests with insertion of additional questions, and deletion of existing questions
-Insertion of tests
- Deletion of tests
- (WIP) Dashboard student list by section

Approved-by: Johndel Villanueva <vjohndel07@yahoo.com>

+++ php/mod_deleteTest.php
@@ -3,28 +3,36 @@ require 'mod_conn.php';
 
 if(isset($_POST['btn-testDelete'])){
     
-    $markedTest = $_POST['testsMark'];
+    if(isset($_POST['testsMark'])){
+        $markedTest = $_POST['testsMark'];
 
-    foreach($markedTest as $marked){
-        
-        // delete question items
-        $q_deleteQuestions = mysqli_query($conn, 
-        "DELETE FROM questions WHERE questions.question_testID = $marked
-        ") or die(mysqli_error($conn));
-        
-        // delete test
-        $q_deleteTest = mysqli_query($conn, 
-        "DELETE FROM tests WHERE tests.test_ID = $marked
-        ") or die(mysqli_error($conn));
+        foreach($markedTest as $marked){
+            
+            // delete question items
+            $q_deleteQuestions = mysqli_query($conn, 
+            "DELETE FROM questions WHERE questions.question_testID = $marked
+            ") or die(mysqli_error($conn));
+            
+            // delete test
+            $q_deleteTest = mysqli_query($conn, 
+            "DELETE FROM tests WHERE tests.test_ID = $marked
+            ") or die(mysqli_error($conn));
 
+        }
+
+        // alert the user for successful deletion of tests
+        echo 
+        "<script>
+            alert('Marked tests successfully deleted!');
+            window.location.href='../dashboard.php';
+        </script>";
+    }else{
+        echo 
+        "<script>
+            alert('No test/s to delete.');
+            window.location.href='../dashboard.php';
+        </script>";
     }
-
-    // alert the user for successful deletion of tests
-    echo 
-    "<script>
-        alert('Marked tests deleted!');
-        window.location.href='../dashboard.php';
-    </script>";
 
 }
 ?>
